+++ packages/orq-rc/src/orq_poc_python_multi_env_version/models/bulkfileuploadop.py
@@ -81,5 +81,5 @@ class ResponseBody(BaseModel):
 
     file_name: str
 
-    created: Optional[datetime] = dateutil.parser.isoparse("2024-11-24T21:26:36.345Z")
+    created: Optional[datetime] = dateutil.parser.isoparse("2024-11-25T07:27:21.148Z")
     r"""The date and time the resource was created"""