+++ packages/orq-rc/src/orq_poc_python_multi_env_version/sdkconfiguration.py
@@ -26,10 +26,10 @@ class SDKConfiguration:
     server_idx: Optional[int] = 0
     language: str = "python"
     openapi_doc_version: str = "2.0"
-    sdk_version: str = "1.42.0-rc.2"
+    sdk_version: str = "1.42.0-rc.6"
     gen_version: str = "2.461.4"
     user_agent: str = (
-        "speakeasy-sdk/python 1.42.0-rc.2 2.461.4 2.0 orq-poc-python-multi-env-version"
+        "speakeasy-sdk/python 1.42.0-rc.6 2.461.4 2.0 orq-poc-python-multi-env-version"
     )
     retry_config: OptionalNullable[RetryConfig] = Field(default_factory=lambda: UNSET)
     timeout_ms: Optional[int] = None